+++ sklearn/svm/classes.py
@@ -52,6 +52,10 @@ class LinearSVC(BaseLibLinear, ClassifierMixin, CoefSelectTransformerMixin):
         To lessen the effect of regularization on synthetic feature weight
         (and therefore on the intercept) intercept_scaling has to be increased
 
+    C_scale_n_samples : bool
+        Scale C with number of samples. It makes the setting of C independant
+        of the number of samples.
+
     Attributes
     ----------
     `coef_` : array, shape = [n_features] if n_classes == 2 else [n_classes, n_features]
@@ -121,6 +125,10 @@ class SVC(DenseBaseLibSVM, ClassifierMixin):
     cache_size: float, optional
         Specify the size of the kernel cache (in MB)
 
+    C_scale_n_samples : bool
+        Scale C with number of samples. It makes the setting of C independant
+        of the number of samples.
+
     Attributes
     ----------
     `support_` : array-like, shape = [n_SV]
@@ -150,8 +158,8 @@ class SVC(DenseBaseLibSVM, ClassifierMixin):
     >>> from sklearn.svm import SVC
     >>> clf = SVC()
     >>> clf.fit(X, y)
-    SVC(C=1.0, cache_size=200, coef0=0.0, degree=3, gamma=0.5, kernel='rbf',
-      probability=False, shrinking=True, tol=0.001)
+    SVC(C=1.0, C_scale_n_samples=False, cache_size=200, coef0=0.0, degree=3,
+      gamma=0.5, kernel='rbf', probability=False, shrinking=True, tol=0.001)
     >>> print clf.predict([[-0.8, -1]])
     [ 1.]
 
@@ -162,10 +170,11 @@ class SVC(DenseBaseLibSVM, ClassifierMixin):
 
     def __init__(self, C=1.0, kernel='rbf', degree=3, gamma=0.0,
                  coef0=0.0, shrinking=True, probability=False,
-                 tol=1e-3, cache_size=200):
+                 tol=1e-3, cache_size=200, C_scale_n_samples=False):
 
         DenseBaseLibSVM.__init__(self, 'c_svc', kernel, degree, gamma, coef0,
-                                 tol, C, 0., 0., shrinking, probability, cache_size)
+                                 tol, C, 0., 0., shrinking, probability,
+                                 cache_size, C_scale_n_samples)
 
 
 class NuSVC(DenseBaseLibSVM, ClassifierMixin):
@@ -269,7 +278,8 @@ class NuSVC(DenseBaseLibSVM, ClassifierMixin):
                  tol=1e-3, cache_size=200):
 
         DenseBaseLibSVM.__init__(self, 'nu_svc', kernel, degree, gamma,
-                                coef0, tol, 0., nu, 0., shrinking, probability, cache_size)
+                                coef0, tol, 0., nu, 0., shrinking, probability,
+                                cache_size, C_scale_n_samples=None)
 
 
 class SVR(DenseBaseLibSVM, RegressorMixin):
@@ -318,6 +328,10 @@ class SVR(DenseBaseLibSVM, RegressorMixin):
     cache_size: float, optional
         Specify the size of the kernel cache (in MB)
 
+    C_scale_n_samples : bool
+        Scale C with number of samples. It makes the setting of C independant
+        of the number of samples.
+
     Attributes
     ----------
     `support_` : array-like, shape = [n_SV]
@@ -346,8 +360,9 @@ class SVR(DenseBaseLibSVM, RegressorMixin):
     >>> X = np.random.randn(n_samples, n_features)
     >>> clf = SVR(C=1.0, epsilon=0.2)
     >>> clf.fit(X, y)
-    SVR(C=1.0, cache_size=200, coef0=0.0, degree=3, epsilon=0.2, gamma=0.2,
-      kernel='rbf', probability=False, shrinking=True, tol=0.001)
+    SVR(C=1.0, C_scale_n_samples=False, cache_size=200, coef0=0.0, degree=3,
+      epsilon=0.2, gamma=0.2, kernel='rbf', probability=False, shrinking=True,
+      tol=0.001)
 
     See also
     --------
@@ -355,11 +370,11 @@ class SVR(DenseBaseLibSVM, RegressorMixin):
     """
     def __init__(self, kernel='rbf', degree=3, gamma=0.0, coef0=0.0,
                  tol=1e-3, C=1.0, epsilon=0.1, shrinking=True,
-                 probability=False, cache_size=200):
+                 probability=False, cache_size=200, C_scale_n_samples=False):
 
         DenseBaseLibSVM.__init__(self, 'epsilon_svr', kernel, degree, gamma,
                                  coef0, tol, C, 0., epsilon, shrinking,
-                                 probability, cache_size)
+                                 probability, cache_size, C_scale_n_samples)
 
     def fit(self, X, y, sample_weight=None, **params):
         """
@@ -433,6 +448,10 @@ class NuSVR(DenseBaseLibSVM, RegressorMixin):
     cache_size: float, optional
         Specify the size of the kernel cache (in MB)
 
+    C_scale_n_samples : bool
+        Scale C with number of samples. It makes the setting of C independant
+        of the number of samples.
+
     Attributes
     ----------
     `support_` : array-like, shape = [n_SV]
@@ -461,8 +480,9 @@ class NuSVR(DenseBaseLibSVM, RegressorMixin):
     >>> X = np.random.randn(n_samples, n_features)
     >>> clf = NuSVR(C=1.0, nu=0.1)
     >>> clf.fit(X, y)
-    NuSVR(C=1.0, cache_size=200, coef0=0.0, degree=3, gamma=0.2, kernel='rbf',
-       nu=0.1, probability=False, shrinking=True, tol=0.001)
+    NuSVR(C=1.0, C_scale_n_samples=False, cache_size=200, coef0=0.0, degree=3,
+       gamma=0.2, kernel='rbf', nu=0.1, probability=False, shrinking=True,
+       tol=0.001)
 
     See also
     --------
@@ -471,10 +491,12 @@ class NuSVR(DenseBaseLibSVM, RegressorMixin):
 
     def __init__(self, nu=0.5, C=1.0, kernel='rbf', degree=3,
                  gamma=0.0, coef0=0.0, shrinking=True,
-                 probability=False, tol=1e-3, cache_size=200):
+                 probability=False, tol=1e-3, cache_size=200,
+                 C_scale_n_samples=False):
 
         DenseBaseLibSVM.__init__(self, 'nu_svr', kernel, degree, gamma, coef0,
-                                 tol, C, nu, None, shrinking, probability, cache_size)
+                                 tol, C, nu, None, shrinking, probability,
+                                 cache_size, C_scale_n_samples)
 
     def fit(self, X, y, sample_weight=None, **params):
         """
@@ -535,6 +557,11 @@ class OneClassSVM(DenseBaseLibSVM):
     cache_size: float, optional
         Specify the size of the kernel cache (in MB)
 
+    C_scale_n_samples : bool
+        Scale C with number of samples. It makes the setting of C independant
+        of the number of samples.
+
+
     Attributes
     ----------
     `support_` : array-like, shape = [n_SV]
@@ -557,7 +584,8 @@ class OneClassSVM(DenseBaseLibSVM):
     def __init__(self, kernel='rbf', degree=3, gamma=0.0, coef0=0.0,
                  tol=1e-3, nu=0.5, shrinking=True, cache_size=200):
         DenseBaseLibSVM.__init__(self, 'one_class', kernel, degree, gamma,
-                                 coef0, tol, 0., nu, 0., shrinking, False, cache_size)
+                                 coef0, tol, 0., nu, 0., shrinking, False,
+                                 cache_size, C_scale_n_samples=None)
 
     def fit(self, X, class_weight={}, sample_weight=None, **params):
         """